#8.14 Bugfix

+++ src/controllers/videoController.js
@@ -3,39 +3,61 @@ import User from '../models/User';
 
 export const home = async (req, res) => {
   const videos = await Video.find({}).sort({ createdAt: 'desc' });
+
   return res.render('home', { pageTitle: 'Home', videos });
 };
 
 export const watch = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id).populate('owner');
+
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
   }
+
   return res.render('watch', { pageTitle: video.title, video });
 };
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id);
+  const {
+    user: { _id },
+  } = req.session;
+
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
+
   return res.render('edit', { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
-  const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
+  const {
+    user: { _id },
+  } = req.session;
+  const { title, description, hashtags } = req.body;
+
   if (!video) {
     return res.render('404', { pageTitle: 'Video not found.' });
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
+
   await Video.findByIdAndUpdate(id, {
     title,
     description,
     hashtags: Video.formatHashtags(hashtags),
   });
+
   return res.redirect(`/videos/${id}`);
 };
 
@@ -49,6 +71,7 @@ export const postUpload = async (req, res) => {
   } = req.session;
   const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
+
   try {
     const newVideo = await Video.create({
       title,
@@ -71,13 +94,28 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const video = await Video.findById(id);
+  const {
+    user: { _id },
+  } = req.session;
+
+  if (!video) {
+    return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
+
   await Video.findByIdAndDelete(id);
+
   return res.redirect('/');
 };
 
 export const search = async (req, res) => {
   const { keyword } = req.query;
   let videos = [];
+
   if (keyword) {
     videos = await Video.find({
       title: {
@@ -85,5 +123,6 @@ export const search = async (req, res) => {
       },
     });
   }
+
   res.render('search', { pageTitle: 'Search', videos });
 };
